// contracts/typechain-types/InvoiceRegistry.ts
export interface InvoiceRegistryInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "cancelInvoice"
      | "createInvoice"
      | "getInvoice"
      | "invoices"
      | "payInvoice"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic:
      | "InvoiceCancelled"
      | "InvoiceCreated"
      | "InvoicePaid"
  ): EventFragment;

  encodeFunctionData(
    functionFragment: "cancelInvoice",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "createInvoice",
    values: [BytesLike, BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getInvoice",
    values: [BytesLike]
  ): string;
  encodeFunctionData(functionFragment: "invoices", values: [BytesLike]): string;
  encodeFunctionData(
    functionFragment: "payInvoice",
    values: [BytesLike]
  ): string;

  decodeFunctionResult(
    functionFragment: "cancelInvoice",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "createInvoice",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "getInvoice", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "invoices", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "payInvoice", data: BytesLike): Result;
}

export namespace InvoiceCancelledEvent {
  export type InputTuple = [code: BytesLike, creator: AddressLike];
  export type OutputTuple = [code: string, creator: string];
  export interface OutputObject {
    code: string;
    creator: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace InvoiceCreatedEvent {
  export type InputTuple = [
    code: BytesLike,
    creator: AddressLike,
    amount: BigNumberish
  ];
  export type OutputTuple = [code: string, creator: string, amount: bigint];
  export interface OutputObject {
    code: string;
    creator: string;
    amount: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace InvoicePaidEvent {
  export type InputTuple = [
    code: BytesLike,
    payer: AddressLike,
    amount: BigNumberish
  ];
  export type OutputTuple = [code: string, payer: string, amount: bigint];
  export interface OutputObject {
    code: string;
    payer: string;
    amount: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface InvoiceRegistry extends BaseContract {
  connect(runner?: ContractRunner | null): InvoiceRegistry;
  waitForDeployment(): Promise<this>;

  interface: InvoiceRegistryInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  cancelInvoice: TypedContractMethod<[code: BytesLike], [void], "nonpayable">;

  createInvoice: TypedContractMethod<
    [code: BytesLike, amount: BigNumberish],
    [void],
    "nonpayable"
  >;

  getInvoice: TypedContractMethod<
    [code: BytesLike],
    [
      [string, bigint, boolean, boolean, bigint] & {
        creator: string;
        amount: bigint;
        paid: boolean;
        cancelled: boolean;
        createdAt: bigint;
      }
    ],
    "view"
  >;

  invoices: TypedContractMethod<
    [arg0: BytesLike],
    [
      [string, bigint, boolean, boolean, bigint] & {
        creator: string;
        amount: bigint;
        paid: boolean;
        cancelled: boolean;
        createdAt: bigint;
      }
    ],
    "view"
  >;

  payInvoice: TypedContractMethod<[code: BytesLike], [void], "payable">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "cancelInvoice"
  ): TypedContractMethod<[code: BytesLike], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "createInvoice"
  ): TypedContractMethod<
    [code: BytesLike, amount: BigNumberish],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "getInvoice"
  ): TypedContractMethod<
    [code: BytesLike],
    [
      [string, bigint, boolean, boolean, bigint] & {
        creator: string;
        amount: bigint;
        paid: boolean;
        cancelled: boolean;
        createdAt: bigint;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "invoices"
  ): TypedContractMethod<
    [arg0: BytesLike],
    [
      [string, bigint, boolean, boolean, bigint] & {
        creator: string;
        amount: bigint;
        paid: boolean;
        cancelled: boolean;
        createdAt: bigint;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "payInvoice"
  ): TypedContractMethod<[code: BytesLike], [void], "payable">;

  getEvent(
    key: "InvoiceCancelled"
  ): TypedContractEvent<
    InvoiceCancelledEvent.InputTuple,
    InvoiceCancelledEvent.OutputTuple,
    InvoiceCancelledEvent.OutputObject
  >;
  getEvent(
    key: "InvoiceCreated"
  ): TypedContractEvent<
    InvoiceCreatedEvent.InputTuple,
    InvoiceCreatedEvent.OutputTuple,
    InvoiceCreatedEvent.OutputObject
  >;
  getEvent(
    key: "InvoicePaid"
  ): TypedContractEvent<
    InvoicePaidEvent.InputTuple,
    InvoicePaidEvent.OutputTuple,
    InvoicePaidEvent.OutputObject
  >;

  filters: {
    "InvoiceCancelled(bytes32,address)": TypedContractEvent<
      InvoiceCancelledEvent.InputTuple,
      InvoiceCancelledEvent.OutputTuple,
      InvoiceCancelledEvent.OutputObject
    >;
    InvoiceCancelled: TypedContractEvent<
      InvoiceCancelledEvent.InputTuple,
      InvoiceCancelledEvent.OutputTuple,
      InvoiceCancelledEvent.OutputObject
    >;

    "InvoiceCreated(bytes32,address,uint256)": TypedContractEvent<
      InvoiceCreatedEvent.InputTuple,
      InvoiceCreatedEvent.OutputTuple,
      InvoiceCreatedEvent.OutputObject
    >;
    InvoiceCreated: TypedContractEvent<
      InvoiceCreatedEvent.InputTuple,
      InvoiceCreatedEvent.OutputTuple,
      InvoiceCreatedEvent.OutputObject
    >;

    "InvoicePaid(bytes32,address,uint256)": TypedContractEvent<
      InvoicePaidEvent.InputTuple,
      InvoicePaidEvent.OutputTuple,
      InvoicePaidEvent.OutputObject
    >;
    InvoicePaid: TypedContractEvent<
      InvoicePaidEvent.InputTuple,
      InvoicePaidEvent.OutputTuple,
      InvoicePaidEvent.OutputObject
    >;
  };
}